Move AssetHelper into shared crate code.
